refactoring: organized utils
