+++ utils/time/convertDateToWeekDayYear.ts
